Timeline fixes
