+++ arpg-timeline/package.json
@@ -34,7 +34,6 @@
         "next-sanity": "^10.0.7",
         "react": "^19.1.1",
         "react-dom": "^19.1.1",
-        "react-google-charts": "^5.2.1",
         "react-is": "^19.1.1",
         "react-responsive": "^10.0.1",
         "styled-components": "^6.1.19",
