--- arpg-timeline/src/components/Timeline/Timeline.css
@@ -1,61 +1,0 @@
-@reference "../../app/globals.css";
-
-.chart g text {
-    @apply fill-foreground!;
-    @apply font-normal!;
-}
-
-.chart g text {
-    @apply cursor-default;
-    @apply select-none;
-}
-
-.chart .google-visualization-tooltip {
-    @apply bg-transparent;
-    @apply shadow-none;
-    @apply border-none;
-    @apply text-popover-foreground;
-}
-
-.chart g rect {
-    rx: 4px;
-    ry: 4px;
-    opacity: 0.7;
-}
-
-.chart g path[stroke-width="1"][stroke="#151519"] {
-    opacity: 0;
-}
-
-.chart g rect:last-of-type[fill="#054735"] + text,
-.chart g rect:last-of-type[fill="#054161"] + text {
-    pointer-events: none;
-}
-
-.chart text {
-    font-family: "Cinzel" !important;
-    font-size: 0.75rem !important;
-}
-
-.chart [width="3"][fill="#054735"],
-.chart [width="3"][fill="#054161"],
-.chart [width="3"][fill="#054735"] + text,
-.chart [width="3"][fill="#054161"] + text {
-    display: none;
-}
-
-.chart g:not(:last-of-type) rect:last-of-type[fill="#054735"],
-.chart g:not(:last-of-type) rect:last-of-type[fill="#054161"] {
-    width: 1px;
-    height: 2000px;
-    fill: hsl(var(--destructive)) !important;
-    transform: translateY(-1969px);
-    opacity: 50%;
-    pointer-events: none;
-    display: block;
-}
-
-/* Fix for noncustomizable focus outline */
-.chart [stroke="#000000"] {
-    @apply stroke-0;
-}
